+++ src/components/CustomPieChart.js
@@ -7,6 +7,11 @@ const renderCustomizedLabel = ({ cx, cy, midAngle, innerRadius, outerRadius, per
     const x = cx + radius * Math.cos(-midAngle * RADIAN);
     const y = cy + radius * Math.sin(-midAngle * RADIAN);
 
+    console.log((percent * 100).toFixed(0));
+    if ((percent * 100).toFixed(0) == 0) {
+        return (<></>);
+    }
+
     return (
         <text x={x} y={y} fill="white" textAnchor={x > cx ? 'start' : 'end'} dominantBaseline="central">
             {`${(percent * 100).toFixed(0)}%`}